Completed revision 5527. Coverity CID 21823.
git-svn-id: http://svn.libburnia-project.org/libburn/trunk@5548 fff8544e-141b-0410-a922-c1ea1f00b384

+++ libburn/mmc.c
@@ -4330,7 +4330,8 @@ static int interpret_performance(struct burn_drive *d, struct command *c,
 		if (descr_type == 0x03) {
 			exact_bit = !!(pd[8 + i*16] & 2);
 			for (b = 0; b < 4 ; b++) {
-				end_lba     += pd[8 + i*16 +  4 + b]
+				end_lba     += ((unsigned long int)
+						pd[8 + i*16 +  4 + b])
 				               << (24 - 8 * b);
 				read_speed  += pd[8 + i*16 +  8 + b]
 				               << (24 - 8 * b);
